feat(data-table): add useExpandRow composable for managing expandable rows
